pyproject.toml
[project]
name = "entity-matching-capitaliq"
version = "1.0.0"
description = "GenAI-powered entity matching for S&P Capital IQ identity reconciliation"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyspark>=3.4.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "sentence-transformers>=2.2.0",
    "databricks-sdk>=0.18.0",
    "mlflow>=2.9.0",
    "gliner>=0.1.0",
    "faiss-cpu>=1.7.4",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tqdm>=4.65.0",
    "scikit-learn>=1.3.0",
    "pyarrow>=12.0.0",
    "delta-spark>=2.4.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.14.0",
    "datasets>=2.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
